ISSUE #2611 elastic.js moved in handler folder elastic/utils.js deleted and its methods moved in the existing utils.js
